requirements.txt
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-telegram-bot==8.1.1